re-use the api-client stuff in the sync connection
