+++ packages/api-client/src/index.js
@@ -1,1 +1,2 @@
 export { TamanuApi } from './TamanuApi';
+export { extractErrorFromFetchResponse } from './errors';
