+++ packages/facility-server/app/sync/FacilitySyncConnection.js
@@ -1,7 +1,6 @@
 import config from 'config';
 
-import { RemoteCallError } from '@tamanu/errors';
-import { getResponseJsonSafely } from '@tamanu/shared/utils';
+import { extractErrorFromFetchResponse } from '@tamanu/api-client';
 import { log } from '@tamanu/shared/services/logging';
 
 /**
@@ -33,15 +32,7 @@ export class FacilitySyncConnection {
     });
 
     if (!response.ok) {
-      const responseBody = await getResponseJsonSafely(response);
-      const { error } = responseBody;
-
-      const errorMessage = error ? error.message : 'no error message given';
-      const err = new RemoteCallError(
-        `Facility Sync API responded with status code ${response.status} (${errorMessage})`,
-      );
-      err.status = response.status;
-      throw err;
+      throw await extractErrorFromFetchResponse(response, url, log);
     }
 
     return await response.json();
